+++ src/main/java/com/github/gaboso/module/impl/GruntModule.java
@@ -1,7 +1,8 @@
-package com.github.gaboso.module;
+package com.github.gaboso.module.impl;
 
 import com.github.gaboso.format.Formatter;
 import com.github.gaboso.helper.CommandHelper;
+import com.github.gaboso.module.Module;
 
 import java.io.File;
 import java.util.Arrays;
@@ -11,12 +12,10 @@ import java.util.Objects;
  * @since 1.0
  * Grunt Module
  */
-public class GruntModule {
+public class GruntModule implements Module {
 
-    private GruntModule() {
-    }
-
-    public static boolean isProject(File[] listFiles) {
+    @Override
+    public boolean isProject(File[] listFiles) {
         return Arrays.stream(listFiles)
                      .filter(Objects::nonNull)
                      .filter(file -> !file.isDirectory())
@@ -24,7 +23,8 @@ public class GruntModule {
                      .anyMatch("gruntfile.js"::equals);
     }
 
-    public static void executeCommands(String projectName, String projectPath) {
+    @Override
+    public void executeCommands(String projectName, String projectPath) {
         Formatter formatter = new Formatter("Grunt", projectName);
 
         CommandHelper.executeCMD(projectPath, "grunt", formatter);